requirements.txt
mcp
starlette
uvicorn
sse-starlette
stripe
supabase
python-dotenv
gunicorn